+++ src/exception/index.js
@@ -1,0 +1,11 @@
+class Exception {
+  checkErrorFor(errorInstance) {
+    errorInstance.occurError();
+  }
+
+  occurError() {
+    throw new Error('OVERRIDING ERROR');
+  }
+}
+
+module.exports = Exception;
